復元５

+++ simplezakka/src/main/java/com/example/simplezakka/dto/order/CustomerInfo.java
@@ -10,15 +10,14 @@ import jakarta.validation.constraints.Size;
 
 @Data
 @NoArgsConstructor
-@AllArgsConstructor 
+@AllArgsConstructor
 public class CustomerInfo {
-
-    /** 顧客ID (会員の場合に設定、ゲストの場合は0) */
-    private Integer customerId; 
+    
+    private Long customerId;
 
     @NotBlank(message = "氏名は必須です。")
     @Size(max = 100, message = "氏名は100文字以内で入力してください。")
-    private String name; // フロントエンドで姓と名を結合して表示することを想定
+    private String name; 
 
     @NotBlank(message = "メールアドレスは必須です。")
     @Email(message = "有効なメールアドレス形式で入力してください。")
